refactor(node-runner): move files
