+++ .github/workflows/lint.yml
@@ -19,6 +19,8 @@ jobs:
         uses: Swatinem/rust-cache@v2.2.1
         with:
           key: clippy
+      - name: Check format
+        run: cargo fmt --all -- --check
       - name: Clippy
         run: cargo clippy --all-targets -- -D warnings
