+++ crates/cli/Cargo.toml
@@ -13,10 +13,15 @@ path = "src/main.rs"
 [dependencies]
 datagen-rs = { path = "../datagen-rs", features = ["all"], version = "0.1.0" }
 datagen-rs-progress-plugin = { path = "../../plugins/progress-plugin", features = [], default-features = false, version = "0.1.0" }
-datagen-rs-node-plugin = { path = "../../plugins/node-plugin", features = ["nodejs"] }
+datagen-rs-node-runner = { path = "../node-runner", features = ["nodejs"], optional = true }
 clap = { version = "4.4", features = ["derive"] }
 serde_json = { version = "1.0", features = ["preserve_order"] }
 indicatif = "0.17"
 colored = "2.0"
 num-format = "0.4"
 anyhow = "1"
+
+[features]
+node = ["datagen-rs-node-runner"]
+default = ["node"]
+
