+++ crates/node-runner/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "datagen-rs-node-runner"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+nodejs = { version = "0.5.0", features = [], optional = true }
+napi = { path = "../../../napi-rs/crates/napi", version = "2.13.3", features = [
+    "napi8",
+    "deferred_trace",
+    "async",
+    "serde-json"] }
+napi-derive = "2.13.0"
+anyhow = "1"
+serde_json = "1"
+datagen-rs = { path = "../../crates/datagen-rs", version = "0.1.0", features = ["all"] }
+futures = "0.3.28"
+
+[features]
+nodejs = ["dep:nodejs"]
