+++ crates/node-runner/src/classes/schema_path.rs
