+++ crates/node-runner/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod classes;
+#[cfg(feature = "nodejs")]
+pub mod runner;
+pub mod util;
+
+#[macro_use]
+extern crate napi_derive;
