+++ packages/datagen/src/lib.rs
@@ -7,8 +7,8 @@ extern crate napi_derive;
 
 use datagen_rs::schema::schema_definition::Schema;
 use datagen_rs::util::helpers::{generate_random_data, get_schema_value};
-use datagen_rs_node_plugin::classes::node_plugin::NodePlugin;
-use datagen_rs_node_plugin::util::traits::IntoNapiResult;
+use datagen_rs_node_runner::classes::node_plugin::NodePlugin;
+use datagen_rs_node_runner::util::traits::IntoNapiResult;
 use datagen_rs_progress_plugin::{PluginWithSchemaResult, ProgressPlugin};
 use napi::threadsafe_function::ThreadsafeFunction;
 use napi::threadsafe_function::{ErrorStrategy, ThreadsafeFunctionCallMode};
