--- plugins/node-plugin/src/runner/mod.rs
@@ -1,1 +1,0 @@
-pub mod node_runner;
